refactor(connections): move the disconnect ownership oracle out of the handler (#6188)
Pure relocation: the census dataclass and label helpers, the spec census, and the single-lock purge-and-revoke transaction (with the artifact-key and judge logic) move verbatim from the aiohttp handler module into the new kiro_crew/connections/ownership.py, so the destructive-decision logic is owned and unit-tested apart from the HTTP layer. Boundary-crossing names go public (DisconnectScope, remove_provider_entry, open_project_dirs, spec_census, agent_spec_sources, is_scope_label); docstrings and comments moved unchanged. The handler keeps auth, parsing, mint teardown, audit, and response shape, and the entry point signature is identical. Tests are import-path updates only; all 50 disconnect behavior assertions pass unchanged.

Closes #6188
